Modifiche minori

+++ WPFPageSwitch/WPFPageSwitch/Command.cs
@@ -25,7 +25,7 @@ namespace WPFPageSwitch
     abstract class Command
     {
         Log l;
-        static private TcpClient s;
+        static protected TcpClient s;
         static protected bool __logged = false;
         static private bool __connected = false;
         static protected string base_path = Properties.Settings.Default.base_path;
@@ -363,12 +363,14 @@ namespace WPFPageSwitch
             }
             byte[] buffer = new byte[1024];
             int size = 1024;
+            int tot_read = 0;
             try
             {
                 while ((size = data_stream.Read(buffer, 0, size)) != 0)
                 {
                     //Scrivo su un file temporaneo. Se tutto va bene sostituisco quello presente
                     //nella cartella dell'utente
+                    tot_read += size;
                     tmp_file.Write(buffer, 0, size);
                 }
             }
@@ -376,17 +378,30 @@ namespace WPFPageSwitch
             {
                 throw new ServerException(Properties.Messaggi.erroreServer, ServerErrorCode.Default);
             }
-            tmp_file.Close();
-            data_stream.Close();
-            //TODO: dimensione e controllo hash
-
+            finally
+            {
+                tmp_file.Close();
+                data_stream.Close();
+            }
+            
+            if (tot_read != this.dim)
+            {
+                throw new ServerException(Properties.Messaggi.datiInconsistenti, ServerErrorCode.DatiInconsistenti);
+            }
+            SHA256 sha_obj = SHA256Managed.Create();
+            byte[] hash_val;
+            hash_val = sha_obj.ComputeHash(File.Open(tmp_path,FileMode.Open));
+            StringBuilder hex = new StringBuilder(hash_val.Length * 2);
+            foreach (byte b in hash_val)
+                hex.AppendFormat("{0:x2}", b);
+            string sha_reale = hex.ToString();
+            if(sha_reale != sha_contenuto.Trim())
+            {
+                throw new ServerException(Properties.Messaggi.datiInconsistenti, ServerErrorCode.DatiInconsistenti);
+            }
             try
             {
                 File.Delete(path_completo);
-            }
-            catch {; }
-            try
-            {
                 File.Move(tmp_path, path_completo);
             }catch(Exception e)
             {
@@ -402,11 +417,121 @@ namespace WPFPageSwitch
 
     class ComandoAggiornaContenutoFile : Command
     {
+        string path;
+        string nome_file;
+        string path_completo;
+        int dim;
+        string sha_contenuto;
+        DateTime t_creazione;
+        FileStream file = null;
+        const string nome_comando = "UPDATE";
+
+        public ComandoAggiornaContenutoFile(string nome_file, string path, int dim = -1,
+                                DateTime t_modifica = new DateTime(), string sha_contenuto = null
+                                )
+            : base()
+        {
+            FileInfo finfo = null;
+            Log l = Log.getLog();
+            if (!Logged)
+            {
+                throw new ServerException(Properties.Messaggi.nonLoggato, ServerErrorCode.UtenteNonLoggato);
+            }
+            this.path = path;
+            this.nome_file = nome_file;
+            this.path_completo = new StringBuilder(base_path).Append(Path.DirectorySeparatorChar).Append(path).Append(Path.DirectorySeparatorChar).Append(nome_file).ToString();
+
+            try
+            {
+                finfo = new FileInfo(path_completo);
+                if (!finfo.Exists)
+                {
+                    throw new Exception("Il file da inviare non esiste.");
+                }
+            }
+            catch (Exception e)
+            {
+                throw new Exception("Errore nel leggere i parametri del file. Forse i parametri sono sbagliati. " + e.Message);
+            }
+            if (dim < 0)
+                dim = (int)(finfo.Length);
+            if (t_creazione == DateTime.MinValue)
+                t_creazione = finfo.CreationTime;
+
+            file = File.Open(this.path_completo, FileMode.Open);
+            if (sha_contenuto == null)
+            {
+                SHA256 sha_obj = SHA256Managed.Create();
+                byte[] hash_val;
+                hash_val = sha_obj.ComputeHash(this.file);
+                this.sha_contenuto = System.Convert.ToBase64String(hash_val);
+            }
+            this.file.Position = 0;
+        }
         /// <summary>
-        /// Comando che gestisce l'aggiornamento del contenuto di un file
+        /// Comando per la creazione di un nuovo file sul server. Se l'utente ha troppi file, il più
+        /// vecchio tra quelli eliminati viene distrutto. Se non ci sono file eliminati da distruggere
+        /// viene generato un errore.
         /// </summary>
+        /// <exception>CommandExeption con un codice corrispondente all'errore riscontrato</exception>
         public override void esegui()
         {
+            StringBuilder sb = new StringBuilder();
+            sb.Append(nome_comando).Append(Environment.NewLine).
+                Append(nome_file).Append(Environment.NewLine).
+                Append(path).Append(Environment.NewLine).
+                Append(t_creazione.Ticks).Append(Environment.NewLine).
+                Append(sha_contenuto).Append(Environment.NewLine).
+                Append(dim).Append(Environment.NewLine).
+                Append(Environment.NewLine);
+            control_stream_writer.Write(sb.ToString());
+            string response = control_stream_reader.ReadLine();
+            response = response.Trim();
+            CommandErrorCode errorCode = (CommandErrorCode)Int32.Parse(response.Split(' ')[0]); //Extract code from response
+            switch (errorCode)
+            {
+                case CommandErrorCode.OKIntermedio:
+                    try
+                    {
+                        data_stream = CollegamentoDati.getCollegamentoDati(control_stream_reader.ReadLine());
+                    }
+                    catch
+                    {
+                        throw new ServerException(Properties.Messaggi.collegamentoDati,
+                            ServerErrorCode.CollegamentoDatiNonDisponibile);
+                    }
+                    break;
+                case CommandErrorCode.FormatoDatiErrato:
+                    throw new ServerException(Properties.Messaggi.formatoDatiErrato, ServerErrorCode.FormatoDatiErrato);
+                case CommandErrorCode.UtenteNonLoggato:
+                    throw new ServerException(Properties.Messaggi.nonLoggato, ServerErrorCode.UtenteNonLoggato);
+                case CommandErrorCode.FileEsistente:
+                    throw new ServerException(Properties.Messaggi.fileEsistente, ServerErrorCode.FileEsistente);
+                case CommandErrorCode.LimiteFileSuperato:
+                    throw new ServerException(Properties.Messaggi.limiteFileSuperato, ServerErrorCode.LimiteFileSuperato);
+                case CommandErrorCode.DatiIncompleti:
+                default:
+                    throw new ServerException(Properties.Messaggi.erroreServer, ServerErrorCode.Default);
+            }
+            byte[] buffer = new byte[1024];
+            int size = 1024;
+            try
+            {
+                while ((size = file.Read(buffer, 0, size)) != 0)
+                {
+                    data_stream.Write(buffer, 0, size);
+                }
+            }
+            catch
+            {
+                throw new ServerException(Properties.Messaggi.erroreServer, ServerErrorCode.Default);
+            }
+        }
+
+        ~ComandoAggiornaContenutoFile()
+        {
+            data_stream.Close();
+            file.Close();
         }
     }
 
@@ -430,8 +555,16 @@ namespace WPFPageSwitch
 
     class ComandoEsci : Command
     {
+        const string nome_comando = "EXIT"; 
         public override void esegui()
         {
+            StringBuilder sb = new StringBuilder();
+            sb.Append(nome_comando).Append(Environment.NewLine).
+                Append(Environment.NewLine);
+            control_stream_writer.Write(sb.ToString());
+            this.control_stream_reader.Close();
+            this.control_stream_writer.Close();
+           
         }
     }
